Changes to be committed:
	modified:   Grid.java
	modified:   GridSquare.java
	modified:   Map.java
	modified:   Player.java

+++ Grid.java
@@ -1,7 +1,7 @@
 public class Grid
 {
 	public GridSquare[][] Map;
-	public Grid(byte[][] arr, String[][] contents) {
+	public Grid(int[][] arr, String[][] contents) {
 		Map = new GridSquare[3][3];
 		for (int x=0;x<3;x++) {
 			for (int y=0;y<3;y++) {
