+++ GridSquare.java
@@ -1,18 +1,18 @@
 public class GridSquare
 {
 	/* bitmask:
-	 * [0][1][2][3][4][5][6][7]
+	 * [0][1][2][3]
 	 * 0: north
 	 * 1: east
 	 * 2: south
 	 * 3: west
-	 * 4-5: not implemented
+	 * 4-31: not implemented
 	 * */
-	public byte canGoToDirections;
+	public int canGoToDirections;
 
 	public String contains;
 
-	public GridSquare(byte canGoTo, String contents) {
+	public GridSquare(int canGoTo, String contents) {
 		canGoToDirections = canGoTo;
 		contains = contents;
 	}
